requirements.txt
Flask==2.2.2
flask[async]
Werkzeug==2.3.7
gunicorn
azure-identity
azure-storage-blob
azure-storage-queue
python-dotenv